Fix so we use a uint16_t for are stream calculation so we get the expected behavior with regard to wrap. PR-SCTP was would not deliver in some cases at wrap.

+++ KERN/netinet/sctp_indata.c
@@ -5325,7 +5325,7 @@ sctp_kick_prsctp_reorder_queue(struct sctp_tcb *stcb,
 {
 	struct sctp_queued_to_read *ctl, *nctl;
 	struct sctp_association *asoc;
-	int tt;
+	uint16_t tt;
 	
 	asoc = &stcb->asoc;
 	tt = strmin->last_sequence_delivered;
